move selector functions to the cgroup model package
