+++ pkg/security/security_profile/profile/profile_proto_dec_v1.go
@@ -11,10 +11,10 @@ package profile
 import (
 	proto "github.com/DataDog/agent-payload/v5/cws/dumpsv1"
 
+	cgroupModel "github.com/DataDog/datadog-agent/pkg/security/resolvers/cgroup/model"
 	"github.com/DataDog/datadog-agent/pkg/security/secl/model"
 	activity_tree "github.com/DataDog/datadog-agent/pkg/security/security_profile/activity_tree"
 	mtdt "github.com/DataDog/datadog-agent/pkg/security/security_profile/activity_tree/metadata"
-	"github.com/DataDog/datadog-agent/pkg/security/security_profile/profile/selector"
 )
 
 // ProtoToSecurityProfile decodes a Security Profile from its protobuf representation
@@ -24,7 +24,7 @@ func ProtoToSecurityProfile(output *SecurityProfile, pathsReducer *activity_tree
 	}
 
 	output.Metadata = mtdt.ProtoMetadataToMetadata(input.Metadata)
-	output.selector = selector.ProtoToWorkloadSelector(input.Selector)
+	output.selector = cgroupModel.ProtoToWorkloadSelector(input.Selector)
 
 	for key, ctx := range input.ProfileContexts {
 		outCtx := &VersionContext{
