+++ pkg/security/security_profile/profile/profile_proto_enc_v1.go
@@ -11,9 +11,9 @@ package profile
 import (
 	proto "github.com/DataDog/agent-payload/v5/cws/dumpsv1"
 
+	cgroupModel "github.com/DataDog/datadog-agent/pkg/security/resolvers/cgroup/model"
 	activity_tree "github.com/DataDog/datadog-agent/pkg/security/security_profile/activity_tree"
 	mtdt "github.com/DataDog/datadog-agent/pkg/security/security_profile/activity_tree/metadata"
-	"github.com/DataDog/datadog-agent/pkg/security/security_profile/profile/selector"
 )
 
 // SecurityProfileToProto incode a Security Profile to its protobuf representation
@@ -26,7 +26,7 @@ func SecurityProfileToProto(input *SecurityProfile) *proto.SecurityProfile {
 		Metadata:        mtdt.ToProto(&input.Metadata),
 		ProfileContexts: make(map[string]*proto.ProfileContext),
 		Tree:            activity_tree.ToProto(input.ActivityTree),
-		Selector:        selector.WorkloadSelectorToProto(&input.selector),
+		Selector:        cgroupModel.WorkloadSelectorToProto(&input.selector),
 	}
 
 	for key, ctx := range input.versionContexts {
